+++ desktop/applis/apps/bean/bean_games/belotebean/src/test/java/cards/belote/beans/DetailsResultsBeloteBeanTest.java
@@ -89,7 +89,7 @@ public final class DetailsResultsBeloteBeanTest extends BeanBeloteCommonTs {
     private static ResultsBelote results(GameBelote _g, int _user) {
         ResultsBelote res_ = new ResultsBelote();
         res_.setGame(_g);
-        res_.getRes().setUser((byte) _user);
+        res_.getRes().setUser(_user);
         res_.initialize(fourPseudos("0","1","2","3"), new CustList<Longs>());
         StringMap<String> s_ = new StringMap<String>();
         s_.addEntry(EnumCardsExporterUtil.SUITS +Suit.SPADE.getSuitSt(),SPADE);
@@ -114,7 +114,7 @@ public final class DetailsResultsBeloteBeanTest extends BeanBeloteCommonTs {
 
     private static GameBelote game5() {
         RulesBelote rules_ = new RulesBelote();
-        DealBelote deal_ = deal1Loose((byte) 3);
+        DealBelote deal_ = deal1Loose(3);
         GameBelote game_ = new GameBelote(GameType.RANDOM, deal_, rules_);
         bid(game_);
         bid(game_);
@@ -132,7 +132,7 @@ public final class DetailsResultsBeloteBeanTest extends BeanBeloteCommonTs {
         _g.ajouterContrat(new BidBeloteSuit());
     }
 
-    private static DealBelote deal1Loose(byte _dealer) {
+    private static DealBelote deal1Loose(int _dealer) {
         CustList<HandBelote> hands_ = new CustList<HandBelote>();
         HandBelote hand_;
         hand_ = new HandBelote();
@@ -183,7 +183,7 @@ public final class DetailsResultsBeloteBeanTest extends BeanBeloteCommonTs {
     private static GameBelote game1() {
         RulesBelote rules_ = new RulesBelote();
         rules_.getAllowedDeclares().put(DeclaresBelote.HUNDRED, BoolVal.TRUE);
-        DealBelote deal_ = deal1Slam((byte) 3);
+        DealBelote deal_ = deal1Slam(3);
         GameBelote game_ = new GameBelote(GameType.RANDOM, deal_, rules_);
         int first_ = game_.playerAfter(deal_.getDealer());
         BidBeloteSuit bid_;
@@ -244,7 +244,7 @@ public final class DetailsResultsBeloteBeanTest extends BeanBeloteCommonTs {
         _game.ajouterUneCarteDansPliEnCoursJoue(_card);
     }
 
-    private static DealBelote deal1Slam(byte _dealer) {
+    private static DealBelote deal1Slam(int _dealer) {
         CustList<HandBelote> hands_ = new CustList<HandBelote>();
         HandBelote hand_;
         hand_ = new HandBelote();
