+++ desktop/applis/apps/core/games/belote/src/main/java/cards/belote/enumerations/BonusBelote.java
@@ -2,13 +2,13 @@ package cards.belote.enumerations;
 
 public enum BonusBelote {
     LAST_TRICK(10);
-    private final int points;
+    private final long points;
 
     BonusBelote(int _points) {
         points = _points;
     }
 
-    public int getPoints() {
+    public long getPoints() {
         return points;
     }
