+++ desktop/applis/apps/core/games/belote/src/main/java/cards/belote/enumerations/DeclaresBeloteRebelote.java
@@ -2,13 +2,13 @@ package cards.belote.enumerations;
 
 public enum DeclaresBeloteRebelote {
     BELOTE_REBELOTE(20);
-    private final int points;
+    private final long points;
 
     DeclaresBeloteRebelote(int _points) {
         points = _points;
     }
 
-    public int getPoints() {
+    public long getPoints() {
         return points;
     }
